+++ tools/publish/update-packages.ts
@@ -1,38 +1,22 @@
+import { execSync } from 'child_process';
 import * as fs from 'fs';
 import * as path from 'path';
-import * as cliProgress from 'cli-progress';
-import { execSync } from 'child_process';
+import { getProcessType, readJSON } from '../utils/utils';
+import ora = require('ora');
 
-function hasFiles(dir: string, ...names: string[]) {
-  for (const name of names) {
-    const file = path.resolve(dir, name);
-
-    if (!fs.existsSync(file)) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-function getProcessType(dir) {
-  if (hasFiles(dir, 'workspace.json', 'package.json', 'nx.json')) {
-    return 'workspace';
-  } else if (hasFiles(dir, 'angular.json', 'package.json', 'nx.json')) {
-    return 'angular';
-  } else {
-    return null;
-  }
-}
-
+/**
+ * Updates the library packages from the base package
+ * @param {string} lib
+ * @param workspace
+ */
 function updateLibraryPackage(lib: string, workspace: any) {
   const targetPath = path.resolve(workspace.projects[lib].root, 'package.json');
   const sourcePath = './tools/base.package.json';
 
-  writeInFile(targetPath, sourcePath);
+  updatePackageFile(targetPath, sourcePath);
 }
 
-function writeInFile(targetPath: string, sourcePath: string) {
+function updatePackageFile(targetPath: string, sourcePath: string) {
   const targetContent = JSON.parse(fs.readFileSync(targetPath, 'utf8'));
   const sourceContent = JSON.parse(fs.readFileSync(sourcePath, 'utf8'));
 
@@ -55,53 +39,36 @@ export class init {
   public nx: any;
 
   public baseDir: string;
-  private workspace: any;
   public libraries: string[] = [];
 
-  private readJSON(dir: string) {
-    const file = path.resolve(this.baseDir, dir);
-    const json = fs.readFileSync(file, 'utf-8');
-    return JSON.parse(json);
-  }
+  private readonly workspace: any;
 
   constructor() {
     let dir = process.cwd();
     const workspaceType = getProcessType(dir);
     if (workspaceType) {
       this.baseDir = dir;
-      this.nx = this.readJSON('nx.json');
-      this.workspace = this.readJSON(`${workspaceType}.json`);
+      this.nx = readJSON(this.baseDir, 'nx.json');
+      this.workspace = readJSON(this.baseDir, `${workspaceType}.json`);
 
+      // Traverse though workspace and filter for libraries
       Object.keys(this.workspace.projects).forEach((key) => {
         if (this.workspace.projects[key].projectType === 'library') {
           this.libraries.push(key);
         }
       });
 
-      console.log(this.libraries);
+      const spinner = ora();
+      updatePackageFile('./tools/base.package.json', './package.json');
 
-      const progressBar = new cliProgress.SingleBar(
-        {},
-        cliProgress.Presets.shades_classic
-      );
-
-      progressBar.start(this.libraries.length, 0, {
-        message: 'Starting',
-      });
-
-      let n = 1;
-
-      writeInFile('./tools/base.package.json', './package.json');
-
+      console.log('Writing data in to package.json file');
       this.libraries.forEach((lib) => {
-        progressBar.update(n++, {
-          message: 'Updating' + lib,
-        });
+        spinner.start(lib);
         updateLibraryPackage(lib, this.workspace);
+        spinner.succeed();
       });
 
       execSync('nx format:write', { stdio: 'pipe' });
-      progressBar.stop();
     }
   }
 }
